playwright and puppeteer moved to test package
