+++ packages/cli/package.json
@@ -71,6 +71,7 @@
     "@puppeteer/replay": "^2.11.1",
     "@syftdata/codehandler": "file:../codehandler",
     "@syftdata/common": "file:../common",
+    "@syftdata/test": "file:../test",
     "@typescript-eslint/eslint-plugin": "^5.51.0",
     "@typescript-eslint/parser": "^5.51.0",
     "chalk": "^4.1",
