+++ packages/test/package.json
@@ -52,6 +52,7 @@
   },
   "dependencies": {
     "@playwright/test": "^1.32.3",
+    "@puppeteer/replay": "^2.11.1",
     "@syftdata/client": "^0.1.8"
   }
 }
