+++ packages/test/tsconfig.json
@@ -2,7 +2,8 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "outDir": "./lib",
-    "lib": ["ES2015", "dom"]
+    "lib": ["ES2015", "dom"],
+    "skipLibCheck": true,
   },
   "include": [
     "./src"
